Move IPC from battle-engine.js to simulator.js
